Ticket view details and add response button
Added the 'ticket details' panel when clicking on the tickets arrow button.
Added a working response button for customer to reply to agent ticket. it will open a text area that can be wriiten on and then submitted to be saved in the response.txt file.
if a ticket is closed, no close ticket buttons will be available