Working New Ticket Button and Form + Error Message
Error button in the ticket view now will open a window form to fill in details of the new ticket they want to submit and send to staff/agents. It will show error messages if all fields arn't filled, when the tickets are submitted; it is saved into ticket.txt and refreshes the ticket row list.

when the application runs into error, like unable to load information from ticket.txt or response.txt, or scenes failing. A corresponding error alert will pop up to inform the user that the application ran into an error and gives instructions on what to do or what not to do, etc. It will display a log message in the terminal using the logging framework and display where the message came from, the level of severity, the log message attached, and the exception that was caught.

Ticket sorting buttons now only display the status of tickets that belong to the currently logged on user
-the currentLoggedUserID and currentUserID needs to be implemented (ie having the app record which user is currently logged on and using it)