added dashboard and log out button
added a dashboard screen and mapped the it's button.
It has a log out screen that clears the usersessiona and bringst the userback to the log in screen.
it displays the current username and userid of the logged in user at teh bottom left of the dashboard.
the dashboard has a welcome message that displays "Welcome to the JavaFX Customer Service Application"